+++ Aras.VS.MethodPlugin/OpenMethodInVS/OpenInVSConstants.cs
@@ -1,0 +1,23 @@
+﻿//------------------------------------------------------------------------------
+// <copyright file="OpenInVSConstants.cs" company="Aras Corporation">
+//     © 2017-2019 Aras Corporation. All rights reserved.
+// </copyright>
+//------------------------------------------------------------------------------
+
+namespace Aras.VS.MethodPlugin.OpenMethodInVS
+{
+	internal static class OpenInVSConstants
+	{
+		public const string ProtocolName = "openinvs";
+
+		public const string ArasVersionParam = "ArasVersion";
+
+		public const string MethodIdParam = "MethodId";
+
+		public const string ServerUrlParam = "ServerUrl";
+
+		public const string DatabaseParam = "Database";
+
+		public const string UserNameParam = "UserName";
+	}
+}